% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxletEstimator.R
\name{BoxletEstimator}
\alias{BoxletEstimator}
\title{BoxletEstimator}
\usage{
BoxletEstimator(
  spatial_df,
  survey_polygon,
  quad_width,
  quad_height,
  quad_x_spacing,
  quad_y_spacing,
  boxlet_sampling_frame = NULL,
  boxlet_frame_settings = NULL,
  trace = F
)
}
\arguments{
\item{spatial_df}{spatial data frame, with coordinates for midpoints. Attributes of spatial_df, should have coordinates with coordinate reference, along with width and height of each sampling unit.}

\item{survey_polygon}{a spatial polygon that is the same coordinate reference as data.}

\item{quad_width}{width of every sampling unit}

\item{quad_height}{height of every sampling unit}

\item{quad_x_spacing}{distance between sampling units on the x-axis}

\item{quad_y_spacing}{distance between sampling units on the spatial_df-axis}

\item{boxlet_sampling_frame}{a list that is the result of the function BoxletEstimatorSamplingFrame() function, if NULL then we build the frame within this function}

\item{boxlet_frame_settings}{list, if  build_boxlet_sampling_frame = T. then these are the parameters passed to the function see ?BoxletEstimatorSamplingFrame}

\item{trace}{print messages as function progresses}
}
\value{
boxlet info, and variance estimator for total
}
\description{
BoxletEstimator
}
\details{
a functon that replicates the Boxlet variance estimator from \insertCite{fewster2011variance}{systematicsurvey}
Assumes rectangle/square survey region, assumes width and height of sampling units are constant quadrants or transects
}
\references{
\insertAllCited{}
}
