% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxletEstimatorSamplingFrame.R
\name{BoxletEstimatorSamplingFrame}
\alias{BoxletEstimatorSamplingFrame}
\title{BoxletEstimatorSamplingFrame}
\usage{
BoxletEstimatorSamplingFrame(
  survey_polygon,
  quad_width,
  quad_height,
  quad_x_spacing,
  quad_y_spacing,
  boxlet_per_sample_width = 2,
  boxlet_per_sample_height = 2,
  trace = F
)
}
\arguments{
\item{survey_polygon}{a spatial polygon that is the same coordinate reference as data.}

\item{quad_width}{width of every sampling unit}

\item{quad_height}{height of every sampling unit}

\item{quad_x_spacing}{distance between sampling units on the x-axis}

\item{quad_y_spacing}{distance between sampling units on the y-axis}

\item{boxlet_per_sample_width}{is building 'boxlet_indicator_matrix' how many boxlets are in a sampling unit for x-axis, ignored if boxlet_indicator_matrix supplied}

\item{boxlet_per_sample_height}{is building 'boxlet_indicator_matrix' how many boxlets are in a sampling unit for y-axis, ignored if boxlet_indicator_matrix supplied}

\item{trace}{print messages as function progresses}
}
\value{
returns a list containing a fine lattice which represents all the boxlets.
}
\description{
BoxletEstimatorSamplingFrame
}
\details{
an auxillary functon that replicates the Boxlet variance estimator from Fewster 2011
Assumes rectangle/square survey region, assumes width and height of sampling units are constant quadrants or transects
}
