% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoststratifiedNonOverlappingEstimator.R
\name{PoststratifiedNonOverlappingEstimator}
\alias{PoststratifiedNonOverlappingEstimator}
\title{PoststratifiedNonOverlappingEstimator}
\usage{
PoststratifiedNonOverlappingEstimator(
  y,
  sample_area,
  survey_area,
  nx_strat = 2,
  ny_strat = 2
)
}
\arguments{
\item{y}{matrix of data which represents data from an areal view of the survey, use NA's for no data/out of bounds}

\item{sample_area}{area of each sampling unit, either a single value of all the same, otherwise a matrix of equal dimensions}

\item{survey_area}{scalar for total survey area}

\item{nx_strat}{number of sampling units in a stratum on the x-axis}

\item{ny_strat}{number of sampling units in a stratum on the y-axis}
}
\value{
Post-stratified variance estimator for population total based on non-overlapping stratum
}
\description{
PoststratifiedNonOverlappingEstimator
}
\details{
a functon that replicates the variacnce estimator from Lundberg & Strand 2014 eq 7
Post stratified variance estimator minus finite population correction.
}
