% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_helper_funs.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{rmvnorm_prec}
\usage{
rmvnorm_prec(mu, prec, n.sims, random_seed)
}
\arguments{
\item{mu}{vector of MLE both fixed and random effect parameters}

\item{prec}{precision matrix, derived from sdreport(obj, getJointPrecision = T)}

\item{n.sims}{integer number of simulations}

\item{random_seed}{integer seed}
}
\value{
matrix of simulated parameter values with dimensions [n.sims x length(mu)]
}
\description{
rmvnorm_prec
}
\details{
simualte parameters from the joint precision matrix derived from a TMB objects
}
