% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialModelEstimator.R
\name{SpatialModelEstimator}
\alias{SpatialModelEstimator}
\title{SpatialModelEstimator}
\usage{
SpatialModelEstimator(spatial_df, formula, mesh, extrapolation_grid,
  family, link, bias_correct = F, control = list(eval.max = 10000,
  iter.max = 10000), convergence = list(grad_tol = 0.001),
  start_par_list = NULL)
}
\arguments{
\item{spatial_df}{data.frame}

\item{formula}{class(formula)  e.g. formula = y ~ 1}

\item{mesh}{data.frame}

\item{extrapolation_grid}{data.frame we expect columns names "x", "y", "area"}

\item{family}{0 = Poisson, 1 = Negative Binomial, 2 = Gaussian, 3 = Gamma}

\item{link}{link function 0 = log, 1 = logit, 2 = probit, 3 = inverse, 4 = identity}

\item{bias_correct}{<bool> whether bias correction is needed in deriving expectations and standard errors for}

\item{control}{list of settings for nlminb}

\item{convergence}{list of convergence setting to check}

\item{start_par_list}{list of inital values for optimisaton, if you have conversion issues can be useful to look at alternative starting values}
}
\value{
: list of estimated objects.
}
\description{
SpatialModelEstimator
}
\details{
Return Estimates and objects from a model based GMRF approach
}
