% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationAdjusted.R
\name{CorrelationAdjusted}
\alias{CorrelationAdjusted}
\title{CorrelationAdjusted}
\usage{
CorrelationAdjusted(y, sample_area, survey_area)
}
\arguments{
\item{y}{sample data of class "RasterLayer" see an example for how this is created.}

\item{sample_area}{area of each sampling unit which has to be a single value i.e. sampling unit area are all the same}

\item{survey_area}{scalar for survey area}
}
\value{
Spatial autocorrealtion variance estimator for the poopulation total
}
\description{
CorrelationAdjusted
}
\details{
a functon that evaluates the variacnce estimator for population total, with an adjustment based on the spatial autocorrelation, based on Morans I
}
