% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoststratifiedOverlappingEstimator.R
\name{PoststratifiedOverlappingEstimator}
\alias{PoststratifiedOverlappingEstimator}
\title{PoststratifiedOverlappingEstimatorAlt}
\usage{
PoststratifiedOverlappingEstimator(y, nx_strat = 1, ny_strat = 1)
}
\arguments{
\item{y}{matrix of data which represents data from an areal view of the survey, use NA's for no data/out of bounds}

\item{nx_strat}{number of sampling units to the right (x-axis) in each strata}

\item{ny_strat}{number of sampling units in the below (y-axis) in each strata}
}
\value{
Post-stratified variance estimator for the poopulation total based on overlapping strata
}
\description{
PoststratifiedOverlappingEstimatorAlt
}
\details{
a functon that replicates the variacnce estimator from Millar & Olsen 1995
Post stratified variance estimator
}
